+++ Tron/LevelPathfindingComponent.cpp
@@ -120,7 +120,7 @@ std::vector<int> LevelPathfindingComponent::FindPath(int startIdx, int endIdx)
 				break;
 			}
 
-			if (path.size() > m_LevelLayout.lock()->GetGridSize())
+			if (static_cast<int>(path.size()) > m_LevelLayout.lock()->GetGridSize())
 			{
 				path.clear();
 				path.push_back(startIdx);
